Long Term Deposit Prediction/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
catboost
time
scikit-learn==1.4.2